Fixed session handling and added image url to magic entity

+++ src/main/java/org/revature/Alcott_P1_Backend/repository/CustomSessionRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 public interface CustomSessionRepository extends JpaRepository<Session, Integer> {
 
     @Transactional
-    long deleteBysessionId(String sessionId);
+    int deleteBysessionId(String sessionId);
 
     boolean existsBysessionId(String session_id);
